inrange: handle NaN cases
Previously, `inrange(x, NaN, y)` will always return true, which is
neither intuitive nor correct according to the definitions[1,2].
Make a minor alteration to the `inrange` implementation to handle the
NaN case with no overhead and add a test case to cover it.

Also added a benchmark case for `inrange`.  The result of this change
is no difference in benchmark performance beyond noise[3]; the data
shows a very very minor improvement in performance.

[1]: https://github.com/ArashPartow/exprtk/blob/cc1b800c2bd1ac3ac260478c915d2aec6f4eb41c/readme.txt#L314
[2]: https://github.com/ArashPartow/exprtk/blob/cc1b800c2bd1ac3ac260478c915d2aec6f4eb41c/readme.txt#L99
[3]: https://gist.github.com/williamspatrick/e4c1b6e20be602b2aa4f0f34db426ff5

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>